External HDF5 dirs in STATISMO_INCLUDE_DIRS
In statismo-config.cmake, explicitly set HDF5_INCLUDE_DIR and
HDF5_INCLUDE_DIRS instead of  @CMAKE_INSTALL_PREFIX@/include/hdf5 and
@CMAKE_INSTALL_PREFIX@/include/hdf5/cpp. This allows
STATISMO_INCLUDE_DIRS to contain the path of the HDF5 include dirs also
in case of external HDF5 libraries. In the case of HDF5 libraries
installed by Statismo, HDF5_INCLUDE_DIR and HDF5_INCLUDE_DIRS are set to
be the in the install directory of statismo, rather than in the build
directory.