Bugfix: alignment of deformation fields fixed.
The alignment was done using a resample image filter, which does a
coordinate transform of the deformation fields.
This is not the correct behavior. Instead, the coordinate system should
stay the same, but the deformation field has to be manipulated to deform
the reference to a data set of different alignment.