Made tempering function dependent on the reference mesh's center of mass.
So far the tempering function in the example was defined such that it would assign the value
0 for points with x-component larger than zero and 1 otherwise. This would only be applicable for
meshes that are centered on the origin. We made it more general, by using the center of mass for the
definition of the tempering (i.e. we temper values that are larger than
the center of mass).